# counter is now thread safe