warning: I am a rust beginner and implement that in the rust hack & tell... ;)