@tailhook 2016-03-09
Well, from the Rust point of view your code is fine.

But we use "threaded" example for benchmarking. And it's expected for a single counter which is modified by every request to be a bottleneck. It's definitely not expected for multi-threaded application, which could benefit from the parallelism to have such single synchronization place.

Have you tried to run `wrk` or any other http benchmarking tool on the server before and after the change? It's interesting on how it influences, at least for microbenchmarks.