@yanns 2016-03-12
good to know that the code is fine from the Rust point of view :smiley: 

Before:
```
wrk -t8 -c 400 http://localhost:3000/
Running 10s test @ http://localhost:3000/
  8 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     8.07ms    1.54ms  25.34ms   77.40%
    Req/Sec     5.94k   660.86     7.96k    80.25%
  472805 requests in 10.02s, 23.00MB read
  Socket errors: connect 0, read 255, write 0, timeout 0
Requests/sec:  47173.28
Transfer/sec:      2.29MB
```

After:
```
wrk -t8 -c 400 http://localhost:3000/
Running 10s test @ http://localhost:3000/
  8 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     7.92ms    1.37ms  21.04ms   66.57%
    Req/Sec     5.85k   511.69     7.16k    77.25%
  465989 requests in 10.01s, 22.66MB read
  Socket errors: connect 0, read 252, write 0, timeout 0
Requests/sec:  46532.18
Transfer/sec:      2.26MB
```

So there is quite a difference.
If you use this as benchmark, then I'd suggest not to merge this.