@tailhook 2016-03-12
Well, it would be fine if you just copy an example and create a `threaded_with_shared_counter.rs`, or something like that. Benchmarking both is actually interesting, especially for machines with the larger number of cores.