@yanns 2016-03-12
OK, I copied the example and added a comment why this file exist.
Thx for the feedback!