@tailhook 2016-03-12
Well, it looks like you're using old version of the example https://travis-ci.org/tailhook/rotor-http/jobs/115569453#L333 master branch is already changed a bit.