@yanns 2016-03-13
Now it's green on rust stable and rust beta, but still fails with rust nightly.