@tailhook 2016-03-13
Yes. Crash in nightly is not your fault. I've squashed commits and merged in. Thanks!